refactor: reorganise components/ with subdirectories
