+++ app/applications/[cycle]/page.tsx
@@ -1,9 +1,9 @@
 import { auth } from '@/auth'
-import ApplicationTable from '@/components/ApplicationTable'
-import DataUploadDialog from '@/components/DataUploadDialog'
-import { HomepageLinkButton, StatisticsLinkButton } from '@/components/LinkButton'
-import NotFoundPage from '@/components/NotFoundPage'
-import { RoleBadge } from '@/components/RoleBadge'
+import DataUploadDialog from '@/components/dialog/DataUploadDialog'
+import { HomepageLinkButton, StatisticsLinkButton } from '@/components/general/LinkButton'
+import NotFoundPage from '@/components/general/NotFoundPage'
+import { RoleBadge } from '@/components/general/RoleBadge'
+import ApplicationTable from '@/components/table/ApplicationTable'
 import prisma from '@/db'
 import { adminAccess, isSuperUser } from '@/lib/access'
 import { formatCycle } from '@/lib/utils'
