+++ app/page.tsx
@@ -1,9 +1,9 @@
 import { signOutAction } from '@/actions/signOut'
 import { auth } from '@/auth'
-import AdminControlPanel from '@/components/AdminControlPanel'
-import { RoleBadge } from '@/components/RoleBadge'
-import SelectAdmissionsCycle from '@/components/SelectAdmissionsCycle'
-import SignOutButton from '@/components/SignOutButton'
+import { RoleBadge } from '@/components/general/RoleBadge'
+import AdminControlPanel from '@/components/homepage/AdminControlPanel'
+import SelectAdmissionsCycle from '@/components/homepage/SelectAdmissionsCycle'
+import SignOutButton from '@/components/homepage/SignOutButton'
 import { isSuperUser } from '@/lib/access'
 import { allAllowedAdmissionsCycles, getAllAdminAndTutorEmails } from '@/lib/users'
 import { Card, Container, Flex, Heading, Separator } from '@radix-ui/themes'
