+++ components/dialog/CandidateCallout.tsx
