+++ components/dialog/UgTutorDialog.tsx
@@ -1,11 +1,11 @@
-import { ApplicationRow } from '@/components/ApplicationTable'
-import CandidateCallout from '@/components/CandidateCallout'
-import CommentItem from '@/components/CommentItem'
-import Dropdown from '@/components/Dropdown'
-import FormWrapper from '@/components/FormWrapper'
-import GenericDialog from '@/components/GenericDialog'
-import LabelText from '@/components/LabelText'
-import TmuaGradeBox from '@/components/TmuaGradeBox'
+import CandidateCallout from '@/components/dialog/CandidateCallout'
+import CommentItem from '@/components/dialog/CommentItem'
+import FormWrapper from '@/components/dialog/FormWrapper'
+import GenericDialog from '@/components/dialog/GenericDialog'
+import TmuaGradeBox from '@/components/dialog/TmuaGradeBox'
+import Dropdown from '@/components/general/Dropdown'
+import LabelText from '@/components/general/LabelText'
+import { ApplicationRow } from '@/components/table/ApplicationTable'
 import { adminAccess } from '@/lib/access'
 import { insertComment, upsertOutcome } from '@/lib/forms'
 import { FormPassbackState } from '@/lib/types'
